Connection is all done on top of logbook now
