+++ Mountain-UI-Companion/Utils/Extension.swift
@@ -38,7 +38,6 @@ extension UIImage
         
         return scaledImage
     }
-    
 }
 
 extension UIImageView
@@ -115,7 +114,7 @@ extension Logger
     static let editProfileTableViewController = Logger(subsystem: subsystem, category: "EditProfileTableViewController")
     
     static let apollo = Logger(subsystem: subsystem, category: "Apollo")
-        
+    
     static let slopesConnection = Logger(subsystem: subsystem, category: "SlopesConnection")
     
     static let bookmarkManager = Logger(subsystem: subsystem, category: "BookmarkManager")
