+++ Mountain-UI-Companion/Utils/NetworkManager.swift
